Fixes #6050 

This PR implements radix/counting sort for integers.

- [x] Implement radix sort.
- [x] Implement radix argsort.
- [x] Actually use radix sort/argsort.
- [x] Tests
- [x] Benchmarks
- [x] Add optimisations
- [x] Make radix sort default for integers/bools
- [x] Docs

cc: @mattip A review would be nice, along with next steps. Feel free to force-push if need be.