@mattip 2018-12-18
tests and benchmarks come before making radix the default for integers.