@hameerabbasi 2018-12-19
> Will you also support radix sort for floats? One can make this work by deriving keys from the floats, as documented here: https://github.com/eloj/radix-sorting#float-keys

I can do `float32`/`float64` easily but I'm not sure if there are integer equivalents to `complex*`/`float80`/`float128`. Those might require multiple passes and/or copies.