@hameerabbasi 2018-12-24
> You can use the PyArray_ISINTEGER macro.

I still need to check for bool and float/double. Also, I need to decide the exact method based on the exact dtype.