@mattip 2018-12-24
You attach the [sort methods to the dtype](https://github.com/numpy/numpy/blob/master/numpy/core/src/multiarray/arraytypes.c.src#L4276), so [this line](https://github.com/numpy/numpy/blob/master/numpy/core/src/multiarray/item_selection.c#L1124) will return NULL if that dtype cannot do radix sort