@hameerabbasi 2018-12-24
Next, I think, comes attaching `kind='radixsort'` to the `sort` and `argsort` functions, writing tests, benchmarks and docs.

It'd also be nice to know where the default sort is decided, if the benchmarks show an improvement and we want to change it.