@mattip 2018-12-24
> where the default sort is decided

Grep for `PyArray_SortkindConverter` and `NPY_QUICKSORT`. The first is used to convert a string to the proper enum. I think there are some other places that used `NY_QUICKSORT` directly.