@hameerabbasi 2018-12-25
Okay, while this works, it's massively slower than quicksort (by about a factor of 3). I went back to the reference repo: https://github.com/eloj/radix-sorting

And it seems he's doing two things different:

1. Multiple passes for 64-bit and more, along with a bit of OOP to decide the key function. Can be simulated by function pointer, but I highly doubt it's any faster/slower than what we're doing.
2. In his benchmark, he excludes the cost of allocating the auxiliary array.