@mattip 2018-12-25
> allocating the auxiliary array

Can you do this once for each int dtype and store it somewhere?