@hameerabbasi 2018-12-25
> Can you do this once for each int dtype and store it somewhere?

It's the exact same size as the input array, so unfortunately not.