@andravin 2018-12-25
Also, it would be nice to have multi-core support so that all available memory bandwidth can be utilized.

Would also be nice to try 11-bit (2048 entry) counters in radix-sort instead of 8-bit (256 entry). These require 3 passes for 32-bit keys instead of 4, but still fit in L1 cache.

Would be nice to have integer and float merge-sort that use SIMD instructions and compare with radix-sort.

This is a lot of stuff, I am just putting it is on the radar. Cheers.