@hameerabbasi 2018-12-26
> Usually we should avoid using a `malloc` in a numeric function. Often it is better to change the API, as the original implementation you reference does, so that any workspace or output buffer is passed as an argument. The user manages the memory and potentially re-uses buffers instead of allocating them repeatedly.

I'm not sure how much effort this would take, but the user doesn't usually manage buffers in NumPy, the exception being `out` kwargs, which are rarely used. I agree with the mixing sentiment, I was hoping to avoid this myself, but the sorting code is written in a way that would disallow this. The `NOT_USED` pointer is a `PyArray`... which means it can't be used for passing in the auxiliary array. I would need to change all sorts to change this.

> Would be nice to have multi-core support so that all available memory bandwidth can be utilized.

In general, NumPy doesn't use multi-core anywhere in its codebase.

> Would be nice to try 11-bit (2048 entry) counters in radix-sort instead of 8-bit (256 entry). These require 3 passes for 32-bit keys instead of 4, but still fit in L1 cache.

[Apparently, the author of the original repo tried this and found it wasn't all it was hyped up to be.](https://github.com/eloj/radix-sorting#-wider-or-narrower-radix) I'm assuming it has to do with byte alignment being more efficient.

> Would be nice to have integer and float merge-sort that use SIMD instructions and compare with radix-sort.

This would either require different compiler flags or ASM code, which has no precedent in NumPy.

> Also, this is a nice C/C++ idiom for zero-initialization of a heap allocated array:
> 
> `size_t cnt0[256] = { 0 };`

Done!