@jakirkham 2018-12-26
@jschueller probably can give you some tips on how to structure the C code to get the compiler to make use of SIMD instructions.