@andravin 2018-12-26
> > Would be nice to have multi-core support so that all available memory bandwidth can be utilized.
> 
> In general, NumPy doesn't use multi-core anywhere in its codebase.

I see. But NumPy already depends on multi-threaded BLAS libraries. I actually don't see why NumPy would not also enable multi-threading elsewhere using the same OpenMP style environment variables that BLAS libraries use for threading control. Default to 1 thread for backwards compatibility (otherwise users who have already built multi-processing into their applications will be impacted).