@hameerabbasi 2018-12-27
cc @mattip I think this is ready, a review would be awesome. I have to update docs and changelog, but otherwise this is ready.