@hameerabbasi 2018-12-27
> I can do float32/float64 easily but I'm not sure if there are integer equivalents to complex*/float80/float128. Those might require multiple passes and/or copies.

It seems the key-derivation given in the repo wasn't fool-proof. It was causing test failures.