@charris 2019-01-01
It would be good to see the benchmark results for this work.