@hameerabbasi 2019-01-01
I'm attaching a notebook where I run the benchmarks. An estimated 3x increase in sorting performance.

[benchmarks.ipynb.zip](https://github.com/numpy/numpy/files/2719531/benchmarks.ipynb.zip)