@hameerabbasi 2019-01-01
> See also #12418. Are there some common bits dealing with benchmarks and tests that could be shared between these two PRs?

I could set up a benchmarking setup for this that performs benchmarks for all sorts if desired.