@hameerabbasi 2019-03-02
Welp, comparing the benchmarks against `master`, it seems like the Timsort guy has done an amazing job.

`master`

```
[ 62.50%] ··· ======= ======= ================================ ============
                kind   dtype             array_type                        
              ------- ------- -------------------------------- ------------
               merge   int16            ('ordered',)            16.5±0.3μs 
               merge   int16           ('reversed',)            20.1±0.8μs 
               merge   int16            ('uniform',)            16.6±0.7μs 
               merge   int16        ('sorted_block', 10)         231±4μs   
               merge   int16       ('sorted_block', 100)         142±2μs   
               merge   int16       ('sorted_block', 1000)        76.1±2μs  
               merge   int16       ('swapped_pair', 0.01)        143±5μs   
               merge   int16       ('swapped_pair', 0.1)         379±7μs   
               merge   int16       ('swapped_pair', 0.5)         751±30μs  
               merge   int16   ('random_unsorted_area', 0.5)     179±3μs   
               merge   int16   ('random_unsorted_area', 0.1)     41.4±3μs  
               merge   int16   ('random_unsorted_area', 0.01)   18.9±0.3μs 
               merge   int16        ('random_bubble', 1)        18.8±0.5μs 
               merge   int16        ('random_bubble', 5)        26.4±0.4μs 
               merge   int16       ('random_bubble', 10)        40.9±0.5μs 
               merge   int32            ('ordered',)            14.5±0.2μs 
               merge   int32           ('reversed',)            20.3±0.4μs 
               merge   int32            ('uniform',)            14.7±0.6μs 
               merge   int32        ('sorted_block', 10)         210±9μs   
               merge   int32       ('sorted_block', 100)         139±5μs   
               merge   int32       ('sorted_block', 1000)       72.4±0.9μs 
               merge   int32       ('swapped_pair', 0.01)        144±4μs   
               merge   int32       ('swapped_pair', 0.1)         377±10μs  
               merge   int32       ('swapped_pair', 0.5)         778±10μs  
               merge   int32   ('random_unsorted_area', 0.5)     181±5μs   
               merge   int32   ('random_unsorted_area', 0.1)     38.3±3μs  
               merge   int32   ('random_unsorted_area', 0.01)    17.1±1μs  
               merge   int32        ('random_bubble', 1)        16.8±0.8μs 
               merge   int32        ('random_bubble', 5)         23.7±1μs  
               merge   int32       ('random_bubble', 10)         36.9±2μs  
               merge   int64            ('ordered',)            15.3±0.1μs 
               merge   int64           ('reversed',)            20.7±0.5μs 
               merge   int64            ('uniform',)            15.5±0.6μs 
               merge   int64        ('sorted_block', 10)         223±6μs   
               merge   int64       ('sorted_block', 100)        156±0.5μs  
               merge   int64       ('sorted_block', 1000)        84.3±1μs  
               merge   int64       ('swapped_pair', 0.01)        145±1μs   
               merge   int64       ('swapped_pair', 0.1)         378±3μs   
               merge   int64       ('swapped_pair', 0.5)         772±10μs  
               merge   int64   ('random_unsorted_area', 0.5)     183±2μs   
               merge   int64   ('random_unsorted_area', 0.1)    36.1±0.4μs 
               merge   int64   ('random_unsorted_area', 0.01)   17.4±0.3μs 
               merge   int64        ('random_bubble', 1)        18.1±0.4μs 
               merge   int64        ('random_bubble', 5)         25.2±1μs  
               merge   int64       ('random_bubble', 10)         36.3±2μs  
              ======= ======= ================================ ============

[ 75.00%] ··· bench_function_base.Sort.time_sort                              ok
[ 75.00%] ··· ======= ======= ================================ =============
                kind   dtype             array_type                         
              ------- ------- -------------------------------- -------------
               merge   int16            ('ordered',)             11.1±0.2μs 
               merge   int16           ('reversed',)             15.8±0.4μs 
               merge   int16            ('uniform',)            11.1±0.02μs 
               merge   int16        ('sorted_block', 10)          326±7μs   
               merge   int16       ('sorted_block', 100)          258±4μs   
               merge   int16       ('sorted_block', 1000)         137±4μs   
               merge   int16       ('swapped_pair', 0.01)         263±1μs   
               merge   int16       ('swapped_pair', 0.1)          389±9μs   
               merge   int16       ('swapped_pair', 0.5)          519±8μs   
               merge   int16   ('random_unsorted_area', 0.5)      142±3μs   
               merge   int16   ('random_unsorted_area', 0.1)     34.3±0.6μs 
               merge   int16   ('random_unsorted_area', 0.01)    13.2±0.2μs 
               merge   int16        ('random_bubble', 1)         13.4±0.1μs 
               merge   int16        ('random_bubble', 5)         21.7±0.3μs 
               merge   int16       ('random_bubble', 10)         34.7±0.2μs 
               merge   int32            ('ordered',)             12.2±0.8μs 
               merge   int32           ('reversed',)             17.0±0.2μs 
               merge   int32            ('uniform',)             11.9±0.2μs 
               merge   int32        ('sorted_block', 10)          320±8μs   
               merge   int32       ('sorted_block', 100)          267±4μs   
               merge   int32       ('sorted_block', 1000)        143±0.6μs  
               merge   int32       ('swapped_pair', 0.01)         264±5μs   
               merge   int32       ('swapped_pair', 0.1)          413±3μs   
               merge   int32       ('swapped_pair', 0.5)          522±9μs   
               merge   int32   ('random_unsorted_area', 0.5)     140±0.5μs  
               merge   int32   ('random_unsorted_area', 0.1)     35.6±0.7μs 
               merge   int32   ('random_unsorted_area', 0.01)    14.2±0.1μs 
               merge   int32        ('random_bubble', 1)         14.2±0.3μs 
               merge   int32        ('random_bubble', 5)        22.9±0.08μs 
               merge   int32       ('random_bubble', 10)        35.6±0.07μs 
               merge   int64            ('ordered',)             14.1±0.1μs 
               merge   int64           ('reversed',)             19.2±0.3μs 
               merge   int64            ('uniform',)            14.3±0.06μs 
               merge   int64        ('sorted_block', 10)          341±7μs   
               merge   int64       ('sorted_block', 100)          273±2μs   
               merge   int64       ('sorted_block', 1000)         151±3μs   
               merge   int64       ('swapped_pair', 0.01)         272±2μs   
               merge   int64       ('swapped_pair', 0.1)          415±10μs  
               merge   int64       ('swapped_pair', 0.5)          541±4μs   
               merge   int64   ('random_unsorted_area', 0.5)      150±2μs   
               merge   int64   ('random_unsorted_area', 0.1)     39.6±0.5μs 
               merge   int64   ('random_unsorted_area', 0.01)    17.0±0.3μs 
               merge   int64        ('random_bubble', 1)         17.4±0.1μs 
               merge   int64        ('random_bubble', 5)         25.7±0.2μs 
               merge   int64       ('random_bubble', 10)         39.6±0.9μs 
              ======= ======= ================================ =============

[ 87.50%] ··· bench_function_base.Sort.time_sort_inplace                      ok
[ 87.50%] ··· ======= ======= ================================ =============
                kind   dtype             array_type                         
              ------- ------- -------------------------------- -------------
               merge   int16            ('ordered',)             7.37±0.2μs 
               merge   int16           ('reversed',)            7.44±0.08μs 
               merge   int16            ('uniform',)            7.39±0.09μs 
               merge   int16        ('sorted_block', 10)        7.61±0.02μs 
               merge   int16       ('sorted_block', 100)         7.77±0.2μs 
               merge   int16       ('sorted_block', 1000)        7.81±0.2μs 
               merge   int16       ('swapped_pair', 0.01)        7.71±0.1μs 
               merge   int16       ('swapped_pair', 0.1)         7.76±0.1μs 
               merge   int16       ('swapped_pair', 0.5)         7.80±0.2μs 
               merge   int16   ('random_unsorted_area', 0.5)     7.57±0.1μs 
               merge   int16   ('random_unsorted_area', 0.1)    7.53±0.06μs 
               merge   int16   ('random_unsorted_area', 0.01)    7.53±0.1μs 
               merge   int16        ('random_bubble', 1)        7.39±0.03μs 
               merge   int16        ('random_bubble', 5)         7.56±0.2μs 
               merge   int16       ('random_bubble', 10)        7.58±0.05μs 
               merge   int32            ('ordered',)            7.65±0.06μs 
               merge   int32           ('reversed',)            7.62±0.07μs 
               merge   int32            ('uniform',)             7.58±0.1μs 
               merge   int32        ('sorted_block', 10)        7.68±0.03μs 
               merge   int32       ('sorted_block', 100)        7.64±0.04μs 
               merge   int32       ('sorted_block', 1000)       7.55±0.04μs 
               merge   int32       ('swapped_pair', 0.01)       7.75±0.07μs 
               merge   int32       ('swapped_pair', 0.1)        7.78±0.05μs 
               merge   int32       ('swapped_pair', 0.5)        7.82±0.03μs 
               merge   int32   ('random_unsorted_area', 0.5)    7.57±0.03μs 
               merge   int32   ('random_unsorted_area', 0.1)    7.46±0.03μs 
               merge   int32   ('random_unsorted_area', 0.01)    7.55±0.1μs 
               merge   int32        ('random_bubble', 1)        7.58±0.08μs 
               merge   int32        ('random_bubble', 5)         7.65±0.2μs 
               merge   int32       ('random_bubble', 10)        7.56±0.06μs 
               merge   int64            ('ordered',)             7.48±0.2μs 
               merge   int64           ('reversed',)            7.58±0.06μs 
               merge   int64            ('uniform',)             7.56±0.1μs 
               merge   int64        ('sorted_block', 10)        7.90±0.09μs 
               merge   int64       ('sorted_block', 100)         7.71±0.1μs 
               merge   int64       ('sorted_block', 1000)        7.69±0.1μs 
               merge   int64       ('swapped_pair', 0.01)        7.71±0.1μs 
               merge   int64       ('swapped_pair', 0.1)         8.00±0.1μs 
               merge   int64       ('swapped_pair', 0.5)        7.78±0.07μs 
               merge   int64   ('random_unsorted_area', 0.5)    7.60±0.05μs 
               merge   int64   ('random_unsorted_area', 0.1)    7.47±0.03μs 
               merge   int64   ('random_unsorted_area', 0.01)   7.43±0.04μs 
               merge   int64        ('random_bubble', 1)        7.48±0.01μs 
               merge   int64        ('random_bubble', 5)        7.64±0.08μs 
               merge   int64       ('random_bubble', 10)        7.62±0.04μs 
              ======= ======= ================================ =============
```

This branch
```
[ 62.50%] ··· ======= ======= ================================ =============
                kind   dtype             array_type                         
              ------- ------- -------------------------------- -------------
               merge   int16            ('ordered',)            15.9±0.03μs 
               merge   int16           ('reversed',)              143±3μs   
               merge   int16            ('uniform',)             15.8±0.1μs 
               merge   int16        ('sorted_block', 10)          184±2μs   
               merge   int16       ('sorted_block', 100)         170±0.8μs  
               merge   int16       ('sorted_block', 1000)         139±1μs   
               merge   int16       ('swapped_pair', 0.01)         141±1μs   
               merge   int16       ('swapped_pair', 0.1)          151±1μs   
               merge   int16       ('swapped_pair', 0.5)          162±2μs   
               merge   int16   ('random_unsorted_area', 0.5)      145±2μs   
               merge   int16   ('random_unsorted_area', 0.1)     141±0.4μs  
               merge   int16   ('random_unsorted_area', 0.01)    146±0.5μs  
               merge   int16        ('random_bubble', 1)         147±0.9μs  
               merge   int16        ('random_bubble', 5)          144±2μs   
               merge   int16       ('random_bubble', 10)          144±2μs   
               merge   int32            ('ordered',)             13.8±0.2μs 
               merge   int32           ('reversed',)              154±2μs   
               merge   int32            ('uniform',)             14.2±0.2μs 
               merge   int32        ('sorted_block', 10)          201±1μs   
               merge   int32       ('sorted_block', 100)         188±0.8μs  
               merge   int32       ('sorted_block', 1000)         151±2μs   
               merge   int32       ('swapped_pair', 0.01)         155±2μs   
               merge   int32       ('swapped_pair', 0.1)          162±3μs   
               merge   int32       ('swapped_pair', 0.5)          167±2μs   
               merge   int32   ('random_unsorted_area', 0.5)     151±0.8μs  
               merge   int32   ('random_unsorted_area', 0.1)     154±0.4μs  
               merge   int32   ('random_unsorted_area', 0.01)     158±2μs   
               merge   int32        ('random_bubble', 1)         155±0.5μs  
               merge   int32        ('random_bubble', 5)         152±0.9μs  
               merge   int32       ('random_bubble', 10)          156±3μs   
               merge   int64            ('ordered',)             15.1±0.7μs 
               merge   int64           ('reversed',)              190±1μs   
               merge   int64            ('uniform',)             15.0±0.3μs 
               merge   int64        ('sorted_block', 10)          243±2μs   
               merge   int64       ('sorted_block', 100)          227±8μs   
               merge   int64       ('sorted_block', 1000)        187±0.7μs  
               merge   int64       ('swapped_pair', 0.01)        196±0.7μs  
               merge   int64       ('swapped_pair', 0.1)          199±4μs   
               merge   int64       ('swapped_pair', 0.5)          197±5μs   
               merge   int64   ('random_unsorted_area', 0.5)      191±1μs   
               merge   int64   ('random_unsorted_area', 0.1)      199±2μs   
               merge   int64   ('random_unsorted_area', 0.01)     202±2μs   
               merge   int64        ('random_bubble', 1)          200±3μs   
               merge   int64        ('random_bubble', 5)          194±4μs   
               merge   int64       ('random_bubble', 10)          197±3μs   
              ======= ======= ================================ =============

[ 75.00%] ··· bench_function_base.Sort.time_sort                              ok
[ 75.00%] ··· ======= ======= ================================ =============
                kind   dtype             array_type                         
              ------- ------- -------------------------------- -------------
               merge   int16            ('ordered',)             14.8±0.1μs 
               merge   int16           ('reversed',)             66.7±0.7μs 
               merge   int16            ('uniform',)             14.6±0.3μs 
               merge   int16        ('sorted_block', 10)         63.8±0.4μs 
               merge   int16       ('sorted_block', 100)         65.4±0.7μs 
               merge   int16       ('sorted_block', 1000)         68.1±2μs  
               merge   int16       ('swapped_pair', 0.01)         67.8±1μs  
               merge   int16       ('swapped_pair', 0.1)          66.3±2μs  
               merge   int16       ('swapped_pair', 0.5)         64.3±0.2μs 
               merge   int16   ('random_unsorted_area', 0.5)     66.5±0.6μs 
               merge   int16   ('random_unsorted_area', 0.1)      68.5±3μs  
               merge   int16   ('random_unsorted_area', 0.01)     76.6±3μs  
               merge   int16        ('random_bubble', 1)          77.0±2μs  
               merge   int16        ('random_bubble', 5)          72.9±2μs  
               merge   int16       ('random_bubble', 10)          71.9±2μs  
               merge   int32            ('ordered',)            15.3±0.09μs 
               merge   int32           ('reversed',)              119±1μs   
               merge   int32            ('uniform',)             15.1±0.1μs 
               merge   int32        ('sorted_block', 10)         140±0.9μs  
               merge   int32       ('sorted_block', 100)         137±0.7μs  
               merge   int32       ('sorted_block', 1000)         116±1μs   
               merge   int32       ('swapped_pair', 0.01)         118±2μs   
               merge   int32       ('swapped_pair', 0.1)          116±2μs   
               merge   int32       ('swapped_pair', 0.5)          108±3μs   
               merge   int32   ('random_unsorted_area', 0.5)     119±0.3μs  
               merge   int32   ('random_unsorted_area', 0.1)      138±9μs   
               merge   int32   ('random_unsorted_area', 0.01)     124±1μs   
               merge   int32        ('random_bubble', 1)          129±4μs   
               merge   int32        ('random_bubble', 5)          123±1μs   
               merge   int32       ('random_bubble', 10)          137±10μs  
               merge   int64            ('ordered',)            17.2±0.09μs 
               merge   int64           ('reversed',)              186±2μs   
               merge   int64            ('uniform',)             17.3±0.4μs 
               merge   int64        ('sorted_block', 10)          251±10μs  
               merge   int64       ('sorted_block', 100)          232±5μs   
               merge   int64       ('sorted_block', 1000)         185±4μs   
               merge   int64       ('swapped_pair', 0.01)         191±5μs   
               merge   int64       ('swapped_pair', 0.1)         198±0.5μs  
               merge   int64       ('swapped_pair', 0.5)          205±10μs  
               merge   int64   ('random_unsorted_area', 0.5)      186±1μs   
               merge   int64   ('random_unsorted_area', 0.1)     189±0.8μs  
               merge   int64   ('random_unsorted_area', 0.01)     198±9μs   
               merge   int64        ('random_bubble', 1)          201±10μs  
               merge   int64        ('random_bubble', 5)          197±7μs   
               merge   int64       ('random_bubble', 10)          190±7μs   
              ======= ======= ================================ =============

[ 87.50%] ··· bench_function_base.Sort.time_sort_inplace                      ok
[ 87.50%] ··· ======= ======= ================================ =============
                kind   dtype             array_type                         
              ------- ------- -------------------------------- -------------
               merge   int16            ('ordered',)             11.0±0.3μs 
               merge   int16           ('reversed',)            10.5±0.07μs 
…